+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Resume previously authenticated sessions via `AuthProvider.resume()`
+
 ## [2.10.0] - 2026-02-25
 
 ### Added